Joao

+++ BackEnd/service/src/main/java/com/basis/sge/service/servico/EventoServico.java
@@ -5,6 +5,7 @@ import com.basis.sge.service.dominio.Evento;
 import com.basis.sge.service.repositorio.EventoRepositorio;
 import com.basis.sge.service.servico.dto.EventoDTO;
 
+import com.basis.sge.service.servico.dto.TipoEventoDTO;
 import com.basis.sge.service.servico.exception.RegraNegocioException;
 import com.basis.sge.service.servico.mapper.EventoMapper;
 import lombok.RequiredArgsConstructor;
@@ -26,21 +27,21 @@ public class EventoServico {
 
     private final EventoMapper eventoMapper;
 
+
     public List<EventoDTO> listar() {
         List<Evento> listaEvento = eventoRepositorio.findAll();
 
         return eventoMapper.toDto(listaEvento);
     }
 
-
     public EventoDTO obterPorId(Integer id){
         Evento evento = eventoRepositorio.findById(id)
-                .orElseThrow(() -> new RegraNegocioException("Usuario não Existe"));
+                .orElseThrow(() -> new RegraNegocioException("Evento não Existe"));
         return eventoMapper.toDto(evento);
     }
 
-
     public EventoDTO criar(EventoDTO eventoDTO) {
+        validaTitulo(eventoDTO.getTitulo());
         Evento evento = eventoMapper.toEntity(eventoDTO);
         Evento eventoSalvo = eventoRepositorio.save(evento);
 
@@ -48,16 +49,25 @@ public class EventoServico {
     }
 
     public EventoDTO atualizar(EventoDTO eventoDTO) {
-
+        validaTitulo(eventoDTO.getTitulo());
         Evento evento = eventoMapper.toEntity(eventoDTO);
         Evento eventoAtualizado = eventoRepositorio.save(evento);
 
         return eventoMapper.toDto(eventoAtualizado);
     }
 
-
     public void deletar(Integer id) {
+        if(!eventoRepositorio.existsById(id)){
+            throw new RegraNegocioException("Evento Não Existe");
+        }
         eventoRepositorio.deleteById(id);
+    }
+
+    //Verifica se o titulo do evento já existe ou esta vazio, em caso positivo solta a exceção
+    public void validaTitulo(String titulo){
+        if (eventoRepositorio.existsByTitulo(titulo)){
+            throw new RegraNegocioException("Um evento com esse titulo já existe");
+        }
     }
 
 }
